finish folder restructuring
