+++ src/DataPage/Charts.scss
@@ -35,7 +35,7 @@
   }
 
   &__histogram {
-    background-image: url('../../shared/img/axis.svg');
+    background-image: url('../shared/img/axis.svg');
     background-position: center;
     background-repeat: no-repeat;
     background-size: auto;
