+++ src/DataPage/Frequencies.js
