+++ src/DataPage/d3/drawEasyBox.js
@@ -106,9 +106,7 @@ const drawEasyBox = () => {
       .data(data)
       .enter()
       .append('svg:image')
-      .attr('xlink:href', (d) =>
-        require(`../../../shared/img/${d.name}.min.svg`)
-      )
+      .attr('xlink:href', (d) => require(`../../shared/img/${d.name}.min.svg`))
       .attr('x', (d) => x(d.avg) - imgSize / 2)
       .attr('y', (d) => y(d.name) - imgSize / 2)
       .attr('width', imgSize)
