+++ src/DataPage/d3/drawFruitHist.js
