+++ src/index.scss
@@ -38,3 +38,58 @@ body {
     width: 100%;
   }
 }
+
+//
+//
+// Styles for the Data and About Pages
+//
+//
+.secondary {
+  padding: 0 1rem 3rem 1rem;
+
+  // SET WIDTHS AND MARGINS
+  // Can't set on a parent because the graphs are a different width
+  main p,
+  .section__heading {
+    max-width: 65rem;
+    margin-left: auto;
+    margin-right: auto;
+  }
+
+  .section__heading {
+    margin-top: 4rem;
+    margin-bottom: 2rem;
+    font-size: 3rem;
+    text-transform: uppercase;
+    letter-spacing: 2px;
+    transition: text-shadow 0.4s;
+
+    &:hover {
+      text-shadow: 2px 2px 0 rgb(255, 0, 183);
+    }
+
+    @media (max-width: 475px) {
+      font-size: 2rem;
+    }
+  }
+
+  main p {
+    margin-bottom: 2rem;
+    line-height: 1.8;
+    font-size: 1.8rem;
+
+    @media (max-width: 475px) {
+      font-size: 1.6rem;
+    }
+
+    a,
+    a:visited {
+      color: rgb(255, 0, 183);
+      text-decoration: none;
+
+      &:hover {
+        color: rgb(199, 1, 143);
+      }
+    }
+  }
+}
