+++ src/shared/components/Toasts/Toasts.scss
