+++ src/shared/hooks/useManageUserRatings.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import fruitList from '../../shared/Fruit.json';
 import toaster from 'toasted-notes';
 import 'toasted-notes/src/styles.css';
-import '../../components/Toasts.scss';
+import '../components/Toasts/Toasts.scss';
 
 const useManageUserRatings = () => {
   const [ratings, setRatings] = useState(fruitList);
